postgresql

+++ CandyShop/pkg/adding/candy.go
@@ -1,8 +1,8 @@
 package adding
 
 type Candy struct {
-	Id       string  `json:"id",omitempty`
-	Category string  `json:"category"`
-	Name     string  `json:"name"`
-	Price    float32 `json:"price"`
+	Id       string `json:"id",omitempty`
+	Category string `json:"category"`
+	Name     string `json:"name"`
+	Price    int    `json:"price"`
 }
